Generate shorter URI for countcomments.action
When using a site with manu comment count elements, the syntax with an
objectId key for every count element generates a very large URI.

To fix this, a common prefix of the objectIDs is determined, which is
often available. The prefix will be send once with the oipr parameter
and the corresponding postfixes with be a comma-separated list in the
oipf parameter.